fix: normalize display math in FileViewer
